[Feat] LocationManager를 Repository로 변경하고 geocoder를 통한 행정구역명 수집, 수집 권한 이벤트 방출 구현
